requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
openai==1.3.7
Pillow==10.1.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
undetected-chromedriver==3.5.4